Separate AvailableFactsAnalysis from SemaBounds.cpp
